fix(geonames): improve geonames api handling
- set retry on geonames api call timeout
- add token setting for private subscriptions
